spi: phytium: Change bus_num from u16 to int for dynamic numbering
The PCI driver stores -1 in bus_num to request dynamic SPI bus
numbering, but the field is u16, so -1 becomes 65535 — a fixed
invalid bus number instead of the SPI core's dynamic sentinel.
Change the type to int, matching the SPI core's bus_num field.

Fixes: caf2f8ff84f6 ("arm64: spi: add Phytium SPI controller support")

+++ drivers/spi/spi-phytium.h
@@ -191,7 +191,7 @@ struct phytium_spi {
 	u32			max_freq;
 
 	u32			reg_io_width;
-	u16			bus_num;
+	int			bus_num;
 	u16			num_cs;
 	int			*cs;
